requirements.txt
easyocr
opencv-python
matplotlib
fuzzywuzzy
gspread
oauth2client
torch
scikit-learn
pytorch-lightning
requests
Pillow
pandas
emoji
tqdm
schedule
pyautogui
PyGetWindow
mss
python-Levenshtein